Add option to configure launching a web app (closes #265, closes #291)
This option replaces an existing `firefoxpwa.openInExistingWindow` preference to allow further configuration how the web app should be launched when it is already opened.

It is now possible to open a new window (default behaviour), open a new tab in the existing window (behavior of previous preference), replace an existing tab, or just focus the window.