Fix settings not displaying in Firefox 116 (fixes #378)
This should make PWAsForFirefox settings should again on `about:preferences` in Firefox 116. I don't know why this happens, and it seems like the order in which scripts get executed was changed, so it misses the initialization. Because of this, the script now tries to immediately add the new preferences instead of waiting for the initialization.

This is not a perfect fix though. If the user opens `about:preferences` on a non-general category and switches to a general category, settings will still be hidden. This will probably be fixed in the future.