Fix loading tabs in containers (fixes #375)
Known problems: When the links target is set to new windows, opening a website with the assigned default container will immediately close the initial window and open a new one with the correct container. Additionally, some widgets will not work in this new window. It is recommended not to use links target set to new windows and assign a default container at the same time. This will likely be addressed in the future.

Future improvements: When the tabs mode is disabled, there is no indicator for the current container. This may be added in the future.