Add feature to make the runtime immutable
This prevents installing and uninstalling the runtime using commands, and also stops patching the runtime when launching web apps. The runtime can still be patched manually from the extension settings or the patch command if the user has sufficient write permissions to do so.

This is meant for immutable systems like Guix or Nix where runtime might be provided by the package manager and pre-patched. See #204 for more.